cleanup more TODO and safety comments
